+++ tsup.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   entry: { index: "src/index.ts", webhooks: "src/webhooks.ts" },
   format: ["esm", "cjs"],
   dts: true,
-  sourcemap: true,
+  sourcemap: false,
   clean: true,
   splitting: false,
   treeshake: true,